# server/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lextures Public API",
    "description": "Versioned REST API for institutional integrations.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/api/v1" }],
  "tags": [
    { "name": "meta", "description": "OpenAPI and documentation" },
    { "name": "courses", "description": "Course resources" },
    { "name": "enrollments", "description": "Roster data" },
    { "name": "users", "description": "User directory" },
    { "name": "assignments", "description": "Assignments" },
    { "name": "grades", "description": "Posted grades" },
    { "name": "graphql", "description": "GraphQL stub" }
  ],
  "paths": {
    "/courses": {
      "get": {
        "tags": ["courses"],
        "summary": "List accessible courses",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 25 } },
          { "name": "fields[courses]", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Paginated course collection" },
          "401": { "description": "Unauthorized", "content": { "application/problem+json": {} } },
          "403": { "description": "Missing scope", "content": { "application/problem+json": {} } },
          "429": { "description": "Rate limited", "content": { "application/problem+json": {} } }
        }
      }
    },
    "/courses/{id}": {
      "get": {
        "tags": ["courses"],
        "summary": "Get course by UUID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Course" },
          "404": { "description": "Not found", "content": { "application/problem+json": {} } }
        }
      }
    },
    "/courses/{id}/enrollments": {
      "get": {
        "tags": ["enrollments"],
        "summary": "List course enrollments",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 25 } }
        ],
        "responses": {
          "200": { "description": "Paginated enrollment collection" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "Get user by UUID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "User (email omitted without pii:read)" }
        }
      }
    },
    "/assignments": {
      "get": {
        "tags": ["assignments"],
        "summary": "List assignments across accessible courses",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated assignments" } }
      }
    },
    "/grades": {
      "get": {
        "tags": ["grades"],
        "summary": "List posted grades across accessible courses",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated grades" } }
      }
    },
    "/openapi.json": {
      "get": {
        "tags": ["meta"],
        "summary": "OpenAPI 3.1 document",
        "responses": { "200": { "description": "OpenAPI JSON" } }
      }
    },
    "/docs": {
      "get": {
        "tags": ["meta"],
        "summary": "Swagger UI",
        "responses": { "200": { "description": "HTML" } }
      }
    },
    "/redoc": {
      "get": {
        "tags": ["meta"],
        "summary": "ReDoc UI",
        "responses": { "200": { "description": "HTML" } }
      }
    },
    "/graphql": {
      "get": {
        "tags": ["graphql"],
        "summary": "GraphQL schema stub (query-only)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "GraphQL SDL" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "API token or JWT" }
    }
  }
}